Create Contacts Service
This service wil be responsible for actually creating the
contacts in Firebase.
Currently there is a minor bug: a contact can be created with
no info (in blank). I hope this can be fixed in the Firebase
side; if not, some additional logic should be put here to
prevent this behavior.